requirements.txt
# MediaCrawler 核心依赖
playwright>=1.40.0
httpx>=0.25.0
tenacity>=8.0.0
pyhumps>=3.8.0
Pillow>=10.0.0

# 飞书 SDK
lark-oapi>=1.4.0

# 数据处理
pydantic>=2.0
pyyaml>=6.0

# 日志
loguru>=0.7.0

# 其他工具
python-dotenv>=1.0.0
